requirements.txt
# Base dependencies
pygame==2.5.2
python-dotenv==1.0.1

# Development dependencies
# (Install with `pip install -r requirements.txt` if not using Poetry)
pytest==8.3.5
coverage==7.8.0
mypy==1.15.0
ruff==0.11.6
black==25.1.0
pre-commit==4.2.0
